refactor: separate accounts and regions apps
