+++ accounts/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('login/', views.login),
+    path('register/', views.signup),
+    path('check-id/', views.check_username),
+]
