+++ config/settings.py
@@ -44,6 +44,8 @@ INSTALLED_APPS = [
     'rest_framework_simplejwt',  
     'corsheaders',
     'analysis',
+    'accounts',
+    'regions',
 ]
 
 MIDDLEWARE = [
