+++ regions/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class RegionsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'regions'
